social_media_api/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
dj-database-url>=2.2
psycopg2-binary>=2.9
whitenoise>=6.6
Pillow>=10.0
python-dotenv>=1.0